refactor : limitTime 변수 타입을 Integer -> Long으로 변경

+++ src/main/java/com/dlqudtjs/codingbattle/dto/room/requestdto/GameRoomCreateRequestDto.java
@@ -51,7 +51,7 @@ public class GameRoomCreateRequestDto {
     @NotNull
     @Min(10)
     @Max(120)
-    private Integer limitTime;
+    private Long limitTime;
 
     public GameRoom toEntity() {
         return GameRoom.builder()
