+++ src/main/java/com/dlqudtjs/codingbattle/dto/room/responsedto/GameRoomStatusResponseDto.java
@@ -16,5 +16,5 @@ public class GameRoomStatusResponseDto {
     private Integer maxUserCount;
     private ProblemLevelType problemLevel;
     private Integer maxSubmitCount;
-    private Integer limitTime;
+    private Long limitTime;
 }
